pyproject.toml
[tool.poetry]
name = "twitter_analytics"
version = "0.1.0"
description = ""
authors = [
    "Lenar Imamutdinov <lenar.imamutdinov@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.6"
Django = "~2.2"
psycopg2 = "~2.8"
pyyaml = "~5.1"
djangorestframework = "^3.10"
django-filter = "~2.2"
uritemplate = "^3.0.0"
tweepy = "~3.8"

[tool.poetry.dev-dependencies]
pytest-django = "^3.5"
factory-boy = "^2.11"
flake8 = "^3.7"
pytest-flake8 = "^1.0"
isort = "^4.3"
pytest-isort = "^0.3"
pytest-pythonpath = "^0.7"
pytest-cov = "^2.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
